--- src/main/java/racingcar/Application.java
@@ -1,7 +1,0 @@
-package racingcar;
-
-public class Application {
-	public static void main(String[] args) {
-		Game.runGame();
-	}
-}
