+++ src/main/java/racingcar/domain/GenerateRandomNo.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static camp.nextstep.edu.missionutils.Randoms.*;
