+++ src/test/java/racingcar/CarTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.AssertionsForClassTypes.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Car;
+
 public class CarTest {
 	@Test
 	@DisplayName("randNo >= 4 이면 true")
